% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_bw_gain.R
\name{ES_bw_gain}
\alias{ES_bw_gain}
\title{ES Weight Gain Calculator}
\usage{
ES_bw_gain(
  df,
  sex = "Sex",
  bwp2 = "BWP2",
  bwp9 = "BWP9",
  nest = "Dam",
  pupID = "PupID",
  condition = "Condition",
  delta_name = "BW_gain",
  analysis_mode = c("pups", "nest")
)
}
\arguments{
\item{df}{Dataframe with body weights}

\item{sex}{Name of column with pup sex info, default is "Sex"}

\item{bwp2}{Name of column with P2 weights, default is "BWP2"}

\item{bwp9}{Name of column with P9 weights, default is "BWP9"}

\item{nest}{Name of column with dam/nest info, default is "Dam"}

\item{pupID}{Name of column with pup IDs, default is "PupID"}

\item{condition}{Name of column with condition info, default is "Condition"}

\item{delta_name}{Name to use for column with body weight gain, default is "BW_gain"}

\item{analysis_mode}{Whether to perform this on a pup or nest level}
}
\value{
Returns a dataframe of P9 - P2 body weights either at the pup or nest level.
}
\description{
This function can be used to calculate pup weight gain in early-life stress (ES) experiments, by subtracting average postnatal day (P)2 weights from the final P9 weights, treating male and female pups separately.
}
\examples{
#generate test column with simulated values
test_df <- data.frame(PupID = rep("Test",10),Condition = rep("CTL", 10), BWP2 = runif(10, 1.5, 3.5), BWP9 = runif(10, 4, 5.2), Sex = c(rep("m", 5),  rep("f", 5)), Dam = c(rep("F01",3), rep("F02",3), rep("F01",2), rep("F02",2)))
ES_bw_gain(test_df) #generates a column BW_gain of P9 - avgP2 weights per pup


}
