% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk_4ways.R
\name{jk_4wayplot}
\alias{jk_4wayplot}
\title{JK Fourway Plot}
\usage{
jk_4wayplot(
  dataframe1,
  dataframe2,
  fc_col = "log2FC",
  fdr_col = "FDR",
  fc_thres = 0.01,
  fdr_thres = 0.05,
  plot_fourway = TRUE,
  genenames = TRUE,
  label_shared_sig = FALSE,
  color_mode = "quadrants",
  symmetry = TRUE,
  color_axes = TRUE,
  color_x_up = "#4d528f",
  color_x_down = "#a3a1c3",
  color_y_up = "#de5b6c",
  color_y_down = "#f3acb0",
  color_x = "#9A9ABC",
  color_y = "#F6A6A9",
  color_both = "black",
  color_ignore = "grey",
  max_overlaps = 20,
  contrast_1 = "Contrast1_Name",
  contrast_2 = "Contrast2_Name",
  label_size = 2.5,
  right_text = "test1_up",
  left_text = "test1_down",
  up_text = "test2_up",
  down_text = "test2_down",
  segment_trans = 0.3,
  unity_alpha = 1,
  aspect = 12/16,
  point_size = 2,
  plot_title = "Significantly altered genes/proteins in both contrasts"
)
}
\arguments{
\item{dataframe1}{first dataframe, will correspond to "X" axis of four-way}

\item{dataframe2}{second dataframe, will correspond to "Y" axis of four-way}

\item{fc_col}{column with fold change info to be plotted, usually Log2-FC or Log-FC}

\item{fdr_col}{column with p-value/q-value/FDR value info to be plotted}

\item{fc_thres}{threshold of fold change to consider. Default is 0 because preference is to still plot genes below the FC threshold as gray.}

\item{fdr_thres}{threshold of p/q/FDR value to plot, default 0.05}

\item{plot_fourway}{boolean of whether to return the fourway plot, or simply return the final dataframe generate}

\item{genenames}{boolean of whether gene names should be plotted on fourway}

\item{label_shared_sig}{boolean of whether labeled genes are those unique to each contrast (by default, obtained when this param is FALSE), or if shared sig genes are labeled (occurs if this param == TRUE)}

\item{color_mode}{affects if points are colored by either "overlap" (i.e., shared or unique between the two contrasts; default) or "quadrants" (includes direction per contrast as well)}

\item{symmetry}{boolean of whether the coordinates should resemble a square (i.e., same values above and below 0 per axis, obtained when param is TRUE) or not}

\item{color_axes}{boolean, default TRUE, of whether axes should be colored or not; uses the same colors used in color_mode quadrant}

\item{color_x_up}{for color_mode "quadrant", color for points defined as "up" in dataframe1}

\item{color_x_down}{for color_mode "quadrant", color for points defined as "down" in dataframe1}

\item{color_y_up}{for color_mode "quadrant", color for points defined as "up" in dataframe2}

\item{color_y_down}{for color_mode "quadrant", color for points defined as "down" in dataframe2}

\item{color_x}{for color_mode "overlap", color for points in dataframe 1}

\item{color_y}{for color_mode "overlap", color for points in dataframe 2}

\item{color_both}{for either color mode, sets color for genes significant in both dataframes}

\item{color_ignore}{for either color mode, sets color for genes to be ignored (usually because they pass significance but are below FC thres)}

\item{max_overlaps}{setting for gene name labeling passed to geom_text_repel; higher value will allow for more genes to be plotted}

\item{contrast_1}{name of the contrast in dataframe1, x axis label}

\item{contrast_2}{name of the contrast in dataframe2, y axis label}

\item{label_size}{font size for gene name labels}

\item{right_text}{= "test1_up",}

\item{left_text}{= "test1_down",}

\item{up_text}{= "test2_up",}

\item{down_text}{text to  "test2_down",}

\item{segment_trans}{transparency of the line that connects gene label to the plot of the gene, default is 0.3 (with 0 being fully transparent, and 1 fully opaque)}

\item{unity_alpha}{=1,}

\item{aspect}{aspect ratio of plot, default is 12/16,}

\item{point_size}{size of points when plotted, default 2}

\item{plot_title}{the title for the plot}
}
\value{
A data frame with shared significant genes/proteins across two contrasts, ready for plotting 4-ways
}
\description{
JK Fourway Plot
}
\examples{
jk_4wayplot(df1, df2, fc_col = "log2FC", fdr_col = "qVal", genenames=FALSE) #basic fourway plot from two DE tables, no genes labeled


}
