% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk_4ways.R
\name{jk_sharedDE}
\alias{jk_sharedDE}
\title{JK Fourway Dataframe}
\usage{
jk_sharedDE(
  dataframe1,
  dataframe2,
  gene_col = "Gene",
  fc_col = "Log2FC",
  fdr_col = "FDR",
  fdr_thres = 0.05,
  fc_thres = 0
)
}
\arguments{
\item{dataframe1}{first dataframe, will correspond to "X" axis of four-way}

\item{dataframe2}{second dataframe, will correspond to "Y" axis of four-way}

\item{fc_col}{column with fold change info to be plotted, usually Log2-FC or Log-FC}

\item{fdr_col}{column with p-value/q-value/FDR value info to be plotted}

\item{fdr_thres}{threshold of p/q/FDR value to plot, default 0.05}

\item{fc_thres}{threshold of fold change to consider. Default is 0 because preference is to still plot genes below the FC threshold as gray.}
}
\value{
A data frame with shared significant genes/proteins across two contrasts, ready for plotting 4-ways
}
\description{
JK Fourway Dataframe
}
\examples{
jk_sharedDE(testvol, fc_col = "log2FC", fdr_col = "qVal") #basic volcano plot using dataframe "testvol"
jk_sharedDE(testvol, fc_col = "log2FC", fdr_col = "qVal", genenames = FALSE) #basic volcano plot with unlabeled points


}
