% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk_volcano.R
\name{jk_volcano}
\alias{jk_volcano}
\title{JK Volcano Plot v1}
\usage{
jk_volcano(
  df,
  gene_col = "Gene",
  fc_col,
  fdr_col,
  fc_thres = 0.1,
  fdr_thres = 0.05,
  contrast_name = "Insert Contrast Here",
  FC_log = 2,
  aspect = 1,
  color_left = "#de5b6c",
  color_right = "#4d528f",
  color_nonsig = "grey",
  point_size = 1.25,
  label_size = 2.5,
  xmargin = 0.1,
  genenames = TRUE,
  max_overlap = 10,
  segment_trans = 0.3,
  right_text = "Contrast_Rt",
  left_text = "Contrast_Lt",
  contrast_text_size = 4
)
}
\arguments{
\item{df}{dataframe to be plotted}

\item{gene_col}{column with Gene symbols}

\item{fc_col}{column with fold change info to be plotted, usually Log2-FC or Log-FC}

\item{fdr_col}{column with p-value/q-value/FDR value info to be plotted}

\item{fc_thres}{threshold of fold change to plot}

\item{fdr_thres}{threshold of p/q/FDR value to plot, default 0.05}

\item{contrast_name}{optional: name of contrast (used in title)}

\item{FC_log}{log base of the foldchange column, used in the x-axis plot (default is 2)}

\item{aspect}{aspect ratio to use, can be expressed as decimal or fraction; default is 1}

\item{color_left}{color of dots on the left side of the volcano plot}

\item{color_right}{color of dots on the right side of the volcano plot}

\item{color_nonsig}{color of filtered out dots on the volcano plot}

\item{point_size}{size of dots in plot}

\item{label_size}{size of labels in plot}

\item{xmargin}{added to the color}

\item{genenames}{TRUE/FALSE, whether gene/protein names should be displayed next to dots, default is TRUE}

\item{max_overlap}{Maximum number of gene names to try to plot, based on geom_text_repel}

\item{segment_trans}{Transparency of line that connects gene names to dots, default 0.3, based on geom_text_repel}

\item{right_text}{Show text on lower right part of plot, uses "color_right"}

\item{left_text}{Show text on lower left part of plot, uses "color_left"}

\item{contrast_text_size}{size of text from 'right_text' and 'left_text'}
}
\value{
A volcano plot showing significantly up/downregulated genes or proteins
}
\description{
JK Volcano Plot v1
}
\examples{
jk_volcano(testvol, fc_col = "log2FC", fdr_col = "qVal", genenames = FALSE) #basic volcano plot with unlabeled points
}
